requirements.txt
# Core
anthropic>=0.40.0
langgraph>=0.2.0
langchain-anthropic>=0.3.0

# Data
yfinance>=0.2.54
pandas>=2.2.0
numpy>=1.26.0
requests>=2.31.0
feedparser>=6.0.11
beautifulsoup4>=4.12.0
lxml>=5.2.0

# Config / storage
pyyaml>=6.0.1
boto3>=1.34.0
pydantic>=2.7.0

# Scheduling / calendar
exchange-calendars>=4.5.0
pytz>=2024.1

# DB
# SQLite is stdlib; no extra package needed

# Dev / testing
pytest>=8.2.0
pytest-asyncio>=0.23.0
python-dotenv>=1.0.1
